add crud Medical Record

+++ src/main/java/likelion/edu/vn/health_care/service/BaseService.java
@@ -7,5 +7,5 @@ public interface BaseService<T> {
     T update(T t);
     void delete(T t);
     Iterable<T> findAll();
-    Optional<T> findById(Long id);
+    Optional<T> findById(Integer id);
 }
