+++ src/main/java/likelion/edu/vn/health_care/service/impl/UserImpl.java
@@ -35,7 +35,7 @@ public class UserImpl implements UserService {
     }
 
     @Override
-    public Optional<UserEntity> findById(Long id) {
+    public Optional<UserEntity> findById(Integer id) {
         return Optional.empty();
     }
 }
